+++ client/console.h
@@ -43,7 +43,7 @@ typedef struct
 
 	int		vislines;
 
-	double	times[NUM_CON_TIMES];	// cls.realtime time the line was generated
+	float	times[NUM_CON_TIMES];	// cls.realtime time the line was generated
 								// for transparent notify lines
 } console_t;
